• Adding the SWC completion support developed by @lucasdupin
This is the first pass at integration, it requires the env vars

    TM_FLEX_PATH
    TM_PROJECT_DIRECTORY

to be set. Swc detection is also very simple, it expects swcs to be in 'lib' or 'lib/name' and won't dig any deeper and will need some polish.

If you have a lot of swc's referenced then the first time you autocomplete there's likely to be a hang. To see the generated stub classes look in `/tmp/tmas3/swcs`.

Thanks to Lucas - this is an amazing feature TM has needed for a long time!